fixed issues with babel and webpack; removed backbone to use globals
