+++ assets/js/admin/collections/colorSettings.js
@@ -1,7 +1,6 @@
-import { Collection } from 'backbone';
 import ColorSetting from '../models/colorSetting';
 
-const ColorSettings = Collection.extend({
+const ColorSettings = Backbone.Collection.extend({
     model: ColorSetting,
 });
