+++ package.json
@@ -22,8 +22,12 @@
         "webpack-dev-server": "^4.15.1"
     },
     "dependencies": {
-        "backbone": "^1.5.0",
+        "backbone": "1.5.0",
+        "jquery": "^3.7.1",
+        "jquery-ui-dist": "^1.13.2",
         "react": "^18.2.0",
-        "react-dom": "^18.2.0"
+        "react-dom": "^18.2.0",
+        "lodash": "4.17.21",
+        "underscore": "1.13.6"
     }
 }
